Moved legacy backfill to it's own function.
